+++ AmazonChimeSDK/AmazonChimeSDK/audiovideo/video/backgroundfilter/SegmentationProcessor.h
@@ -13,7 +13,7 @@
 
 @protocol SegmentationProcessor
 
-- (BOOL) initializeWithHeight:(NSInteger)height width:(NSInteger)width channels:(NSInteger)channels;
+- (BOOL) initialize:(NSInteger)height width:(NSInteger)width channels:(NSInteger)channels;
 
 - (BOOL) predict;